Add multi-process management for Windows CE execution
Introduces concurrent and sequential execution of multiple Windows CE binaries, along with process management features: stopping all processes, showing running processes, and a process monitor stub. Updates UI and logic to support these features, adds process tracking to WindowsCEExecutor, and improves execution result reporting.

+++ InstructionTranslator.cs
@@ -138,7 +138,7 @@ namespace ProcessorEmulator
             {
                 var immediate = operand2 & 0xFF;
                 var rotate = (operand2 >> 8) & 0xF;
-                var value = RotateRight(immediate, rotate * 2);
+                var value = RotateRight(immediate, (int)(rotate * 2));
                 cpu.Registers[rd] = value;
             }
